services/Auth Service/pyproject.toml
[build-system]
requires = ["setuptools==68.2.2", "wheel==0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-service"
version = "0.1.0"
description = "FlashSupport centralized authentication and trust service"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.0",
  "sqlalchemy==2.0.0",
  "psycopg[binary]==3.2.13",
  "pydantic==2.8.0",
  "PyYAML==6.0.0",
  "PyJWT==2.8.0",
  "cryptography==43.0.1",
  "bcrypt==4.1.3",
]

[project.optional-dependencies]
test = [
  "pytest==8.3.0",
  "httpx==0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["domain", "main", "models", "routers", "services"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["infrastructure*"]
